+ Hid the create project type and theme from create new manu + Fixed bamboo tests
git-svn-id: http://agilefant.googlecode.com/svn/branches/Monza_restructure@2415 29784df9-99d5-7e28-3e6c-ac1830307ffc

+++ test/fi/hut/soberit/agilefant/business/IterationBusinessTest.java
@@ -186,8 +186,8 @@ public class IterationBusinessTest {
         
         IterationMetrics actualMetrics = iterationBusiness.getIterationMetrics(iteration); 
         
-        assertEquals(0, actualMetrics.getEffortLeft().getMinorUnits());
-        assertEquals(0, actualMetrics.getOriginalEstimate().getMinorUnits());
+        assertEquals(0, actualMetrics.getEffortLeft().getMinorUnits().longValue());
+        assertEquals(0, actualMetrics.getOriginalEstimate().getMinorUnits().longValue());
         
         verify(iterationHistoryEntryBusiness);
     }
